requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
huggingface_hub
datasets
python-dotenv
tqdm
torch
